"Updated login and account setting pages, added password validation and update functionality in JS and PHP"
This diff includes changes to the login.php file where the column name for the password field in the database table is changed to 'Password' from 'pass'. It also includes changes to the account-setting.php file where the HTML structure is updated to include a section for changing the user's password. The corresponding JS file, main.js, is also updated to handle the password change functionality, including validation of the new password and disabling the save button if the new password is too short. The update.php file is updated to handle the password change request from the client-side and update the password in the database. The init.php file is updated to include a new constant for the reset password page. The general-info.php file is also updated to include a success message paragraph.