"Added area column to property table, formatted property price, and updated related files"
This commit adds a new column 'area' to the 'property' table in the database. It also modifies the 'backend.php' file to format the 'propertyPrice' field using the 'number\_format' function before inserting it into the database. Additionally, the 'My Properties.php' and 'create listing.php' files have been updated to display the new 'area' column in the table. Finally, the 'test.php' file has been updated to demonstrate the use of the 'number\_format' function.