Changes in ar/css/main.css
Two new properties were added to the CSS rule for input[type="number"]::-webkit-outer-spin-button: flex-wrap: wrap and gap: 10px. This change will affect the layout and spacing of the spin button in number input fields.
A new media query was added for screen widths up to 786px (small screens). This query introduces several new CSS rules for the .update-message class:
.update-message is positioned 200px above its original position and 2% from the right edge of the screen.
.update-message::after is positioned 15% from the top and 5% from the right edge of the screen.
.update-message.show has an opacity of 1, is positioned 100px below its original position, and has a transition effect for transform and opacity.
.update-message.hide has an opacity of 0, is positioned 200px above its original position, and has a transition effect for transform and opacity.
These changes will affect the layout and behavior of the .update-message element on small screens.

Changes in user/ar/css/user.css

A single property was changed in the CSS rule for button.disabled: align-items was changed from center to end. This change will affect the alignment of content within disabled buttons.
Changes in visitor_count.txt

The contents of the file were updated from 457 to 494. This file likely tracks the number of visitors to a website or application, and the update reflects an increase of 37 visitors